Update Javad_Test_001.CR6
Changes the Begin Time and End Time of the "TimeIsBetween" instruction; originally the instruction was thought to operate with a time datum of 0 seconds equal to midnight, Sunday morning. This is not the case. The time datum of 0 seconds is equal to the time the program was "last compiled" in the logger; the first instance SW12-2 will become "True" is 604800 seconds (one week) after the program was "last compiled" in the logger.